Refactoring a bit, putting the generic stuff in a common library, should add tests, but haven't
