+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/ostelco-core
+
+go 1.12
